data_collector/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
netifaces==0.10.9
psutil==5.7.0
py-dmidecode==0.1.0
requests==2.23.0
urllib3==1.25.9
